predict.py
#!/usr/bin/env python3
"""
Medical Image Analysis - Inference Script

This script handles inference on new medical images using trained models.
Supports both single image and batch processing modes.
"""

import argparse
from pathlib import Path

import torch
import yaml


def parse_arguments():
    """Parse command line arguments for inference configuration."""
    parser = argparse.ArgumentParser(description='Run inference on medical images')
    
    parser.add_argument('--config', type=str, default='config/default_config.yaml',
                       help='Path to configuration file')
    parser.add_argument('--model', type=str, choices=['unet', 'deeplabv3'], 
                       required=True, help='Model architecture to use')
    parser.add_argument('--checkpoint', type=str, required=True,
                       help='Path to model checkpoint')
    parser.add_argument('--input', type=str, required=True,
                       help='Path to input image or directory')
    parser.add_argument('--output-dir', type=str, default='outputs/predictions',
                       help='Output directory for predictions')
    parser.add_argument('--batch-size', type=int, default=16,
                       help='Batch size for inference')
    parser.add_argument('--device', type=str, default='auto',
                       help='Device to use (cuda, cpu, or auto)')
    parser.add_argument('--save-visualizations', action='store_true',
                       help='Save visualization overlays')
    parser.add_argument('--output-format', type=str, choices=['json', 'csv', 'both'],
                       default='json', help='Output format for results')
    
    return parser.parse_args()

def load_config(config_path):
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config

def main():
    """Main inference function."""
    args = parse_arguments()
    
    # Load configuration
    config = load_config(args.config)
    
    # Override config with command line arguments
    if args.model:
        config['model']['architecture'] = args.model
    if args.batch_size:
        config['data']['batch_size'] = args.batch_size
    
    # Set device
    if args.device == 'auto':
        device = 'cuda' if torch.cuda.is_available() else 'cpu'
    else:
        device = args.device
    
    print(f"Starting inference with configuration:")
    print(f"Model: {config['model']['architecture']}")
    print(f"Checkpoint: {args.checkpoint}")
    print(f"Input: {args.input}")
    print(f"Output directory: {args.output_dir}")
    print(f"Device: {device}")
    print(f"Output format: {args.output_format}")
    
    # Create output directories
    Path(args.output_dir).mkdir(parents=True, exist_ok=True)
    
    # TODO: Initialize and run inference pipeline
    print("Inference pipeline will be implemented in subsequent tasks...")

if __name__ == '__main__':
    main()