requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging
pydicom>=2.4.0
SimpleITK>=2.3.0
nibabel>=5.1.0

# Computer Vision and Augmentations
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0
tifffile>=2023.7.10

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Data Handling
pandas>=2.0.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and Logging
PyYAML>=6.0
tensorboard>=2.13.0
tqdm>=4.65.0

# CLI and Utilities
click>=8.1.0
argparse

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0